feat(crons): Show error for invalid schedules
