+++ static/app/views/monitors/components/monitorCreateForm.tsx
@@ -1,4 +1,4 @@
-import {Fragment, useContext, useEffect, useRef} from 'react';
+import {Fragment, useRef} from 'react';
 import {browserHistory} from 'react-router';
 import {css} from '@emotion/react';
 import styled from '@emotion/styled';
@@ -9,54 +9,27 @@ import SelectField from 'sentry/components/forms/fields/selectField';
 import SentryProjectSelectorField from 'sentry/components/forms/fields/sentryProjectSelectorField';
 import TextField from 'sentry/components/forms/fields/textField';
 import Form from 'sentry/components/forms/form';
-import FormContext from 'sentry/components/forms/formContext';
-import FormModel, {FieldValue} from 'sentry/components/forms/model';
+import FormModel from 'sentry/components/forms/model';
 import Panel from 'sentry/components/panels/panel';
 import PanelBody from 'sentry/components/panels/panelBody';
-import Placeholder from 'sentry/components/placeholder';
 import {timezoneOptions} from 'sentry/data/timezones';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {isActiveSuperuser} from 'sentry/utils/isActiveSuperuser';
-import {useApiQuery} from 'sentry/utils/queryClient';
 import commonTheme from 'sentry/utils/theme';
-import {useDimensions} from 'sentry/utils/useDimensions';
 import useOrganization from 'sentry/utils/useOrganization';
 import usePageFilters from 'sentry/utils/usePageFilters';
 import useProjects from 'sentry/utils/useProjects';
 import {normalizeUrl} from 'sentry/utils/withDomainRequired';
+import {MockTimelineVisualization} from 'sentry/views/monitors/components/mockTimelineVisualization';
 import {
   DEFAULT_CRONTAB,
   DEFAULT_MONITOR_TYPE,
   mapMonitorFormErrors,
   transformMonitorFormData,
 } from 'sentry/views/monitors/components/monitorForm';
-import {MockCheckInTimeline} from 'sentry/views/monitors/components/overviewTimeline/checkInTimeline';
-import {
-  GridLineOverlay,
-  GridLineTimeLabels,
-} from 'sentry/views/monitors/components/overviewTimeline/gridLines';
-import {TimelinePlaceholder} from 'sentry/views/monitors/components/overviewTimeline/timelinePlaceholder';
-import {getConfigFromTimeRange} from 'sentry/views/monitors/components/overviewTimeline/utils';
 import {Monitor, ScheduleType} from 'sentry/views/monitors/types';
 import {crontabAsText, getScheduleIntervals} from 'sentry/views/monitors/utils';
-
-const NUM_SAMPLE_TICKS = 9;
-
-interface ScheduleConfig {
-  cronSchedule?: FieldValue;
-  intervalFrequency?: FieldValue;
-  intervalUnit?: FieldValue;
-  scheduleType?: FieldValue;
-}
-
-function isValidConfig(schedule: ScheduleConfig) {
-  const {scheduleType, cronSchedule, intervalFrequency, intervalUnit} = schedule;
-  return !!(
-    (scheduleType === ScheduleType.CRONTAB && cronSchedule) ||
-    (scheduleType === ScheduleType.INTERVAL && intervalFrequency && intervalUnit)
-  );
-}
 
 const DEFAULT_SCHEDULE_CONFIG = {
   scheduleType: 'crontab',
@@ -64,116 +37,6 @@ const DEFAULT_SCHEDULE_CONFIG = {
   intervalFrequency: '1',
   intervalUnit: 'day',
 };
-
-interface Props {
-  schedule: ScheduleConfig;
-}
-
-function MockTimelineVisualization({schedule}: Props) {
-  const {scheduleType, cronSchedule, intervalFrequency, intervalUnit} = schedule;
-  const organization = useOrganization();
-  const {form} = useContext(FormContext);
-
-  const query = {
-    num_ticks: NUM_SAMPLE_TICKS,
-    schedule_type: scheduleType,
-    schedule:
-      scheduleType === 'interval' ? [intervalFrequency, intervalUnit] : cronSchedule,
-  };
-
-  const elementRef = useRef<HTMLDivElement>(null);
-  const {width: timelineWidth} = useDimensions<HTMLDivElement>({elementRef});
-
-  const sampleDataQueryKey = [
-    `/organizations/${organization.slug}/monitors-schedule-data/`,
-    {query},
-  ] as const;
-  const {data, isLoading, isError, error} = useApiQuery<number[]>(sampleDataQueryKey, {
-    staleTime: 0,
-    enabled: isValidConfig(schedule),
-    retry: false,
-  });
-
-  const errorMessage =
-    isError || !isValidConfig(schedule)
-      ? error?.responseJSON?.schedule?.[0] ?? t('Invalid Schedule')
-      : null;
-
-  useEffect(() => {
-    if (!form) {
-      return;
-    }
-
-    if (scheduleType === ScheduleType.INTERVAL) {
-      form.setError('config.schedule.frequency', errorMessage);
-    } else if (scheduleType === ScheduleType.CRONTAB) {
-      form.setError('config.schedule', errorMessage);
-    }
-  }, [errorMessage, form, scheduleType]);
-
-  const mockTimestamps = data?.map(ts => new Date(ts * 1000));
-  const start = mockTimestamps?.[0];
-  const end = mockTimestamps?.[mockTimestamps.length - 1];
-  const timeWindowConfig =
-    start && end ? getConfigFromTimeRange(start, end, timelineWidth) : undefined;
-
-  return (
-    <TimelineContainer>
-      <TimelineWidthTracker ref={elementRef} />
-      {isLoading || !start || !end || !timeWindowConfig || !mockTimestamps ? (
-        <Fragment>
-          <Placeholder height="40px" />
-          {errorMessage ? <Placeholder height="100px" /> : <TimelinePlaceholder />}
-        </Fragment>
-      ) : (
-        <Fragment>
-          <StyledGridLineTimeLabels
-            timeWindowConfig={timeWindowConfig}
-            start={start}
-            end={end}
-            width={timelineWidth}
-          />
-          <StyledGridLineOverlay
-            showCursor={!isLoading}
-            timeWindowConfig={timeWindowConfig}
-            start={start}
-            end={end}
-            width={timelineWidth}
-          />
-          <MockCheckInTimeline
-            width={timelineWidth}
-            mockTimestamps={mockTimestamps.slice(1, mockTimestamps.length - 1)}
-            start={start}
-            end={end}
-            timeWindowConfig={timeWindowConfig}
-          />
-        </Fragment>
-      )}
-    </TimelineContainer>
-  );
-}
-
-const TimelineContainer = styled(Panel)`
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: 40px 100px;
-  align-items: center;
-`;
-
-const StyledGridLineTimeLabels = styled(GridLineTimeLabels)`
-  grid-column: 0;
-`;
-
-const StyledGridLineOverlay = styled(GridLineOverlay)`
-  grid-column: 0;
-`;
-
-const TimelineWidthTracker = styled('div')`
-  position: absolute;
-  width: 100%;
-  grid-row: 1;
-  grid-column: 0;
-`;
 
 export default function MonitorCreateForm() {
   const organization = useOrganization();
